To download file

+++ internal/domain/file.go
@@ -10,7 +10,7 @@ type File struct {
 	ID               uuid.UUID      `json:"id"`
 	Name      		string    `json:"name"`     
 	MimeType  		string    `json:"mimeType"` 
-	Size      		int64     `json:"size"`
+	Size      		int     `json:"size"`
 	Path      		string    `json:"path"`
 	Base64			string 			`json:"base64"`
 	CreatedAt 		time.Time `json:"createdAt"`
